Reword 'hour' requirement for [[HourCycle]]
"[...] [[Pattern]] includes the hour." sounds a bit strange, so I replaced it
with "[...] [[Pattern]] includes an hour component.". UTS 35 uses 'field' to
describe the individual parts of pattern strings, but reusing that term may
cause confusion with the normal use of 'field' within ECMAScript, therefore
I've picked 'component' for now.