Set field widths for date/timeStyle (#52)
* Editorial: Adjust Intl.DateTimeFormat internal slots to include field widths

Change [[DateFormat]] and [[TimeFormat]] to use a record instead of a simple
pattern. That way field widths (e.g. display "hour" as "2-digit") can be stored.

* Editorial: Return a format record from DateTimeStylePattern

The returned record has the same structure as the one returned from
`BasicFormatMatcher` and `BestFitFormatMatcher`.

* Editorial: Merge setting DateTimeFormat internal slots

Unify the steps to initialise the internal slots of a DateTimeFormat object
from the format record.

* Editorial: Remove no longer needed step to set [[HourCycle]].

This now happens as part of the normal hour-cycle computation steps.

* Editorial: Revert moving hour-cycle computation

This change from ECMA-402 is no longer needed.

* Editorial: Modify resolvedOptions to not include date-time components when date/timeStyle is used

The DateTimeFormat's internal slots for date-time compenents are now correctly
filled, so they need to explicitly filtered out in `resolvedOptions()`.

* Editorial: Remove no longer needed description modification

This change from ECMA-402 isn't needed anymore, too.